+++ packages/xellar-kit/src/styles/styled.ts
@@ -1,0 +1,8 @@
+import styledComponents from 'styled-components';
+
+export const styled = (
+  typeof styledComponents.div === 'function'
+    ? styledComponents
+    : //@ts-expect-error False positive
+      styledComponents['default']
+) as typeof styledComponents;