+++ apps/next-app/CHANGELOG.md
@@ -1,5 +1,26 @@
 # example-next-app
 
+## 0.1.18
+
+### Patch Changes
+
+- Updated dependencies [0501ec3]
+  - @xellar/kit@1.1.8
+
+## 0.1.17
+
+### Patch Changes
+
+- Updated dependencies [66b539b]
+  - @xellar/kit@1.1.7
+
+## 0.1.16
+
+### Patch Changes
+
+- Updated dependencies [8412f0e]
+  - @xellar/kit@1.1.6
+
 ## 0.1.15
 
 ### Patch Changes